apps/api/requirements.txt
# SpectraAI API - 2025 Edition
# Modern FastAPI with async patterns, type hints, and professional structure

# Core dependencies
fastapi>=0.110.0
uvicorn[standard]>=0.27.1
pydantic>=2.6.3
pydantic-settings>=2.2.1
python-multipart>=0.0.9

# Database & ORM
prisma>=0.11.0
asyncpg>=0.29.0

# Authentication & Security
clerk-python>=1.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0

# HTTP client
httpx>=0.27.0

# System monitoring
psutil>=5.9.0

# WebSocket & real-time
websockets>=12.0
python-socketio>=5.11.1

# AI/ML
openai>=1.12.0
httpx>=0.27.0
numpy>=1.26.4
torch>=2.2.1

# Development & Testing
pytest>=8.1.1
pytest-asyncio>=0.23.5
pytest-cov>=4.0.0
ruff>=0.3.0
black>=24.2.0
mypy>=1.8.0

# Production
gunicorn>=21.2.0
python-dotenv>=1.0.1

# Utilities
python-slugify>=8.0.4
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dateutil>=2.9.0
